test: trim redundant native protocol coverage
